pyproject.toml
[project]
name = "image-compressor"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = ["pip>=25.2", "autopep8>=2.3.2", "numpy>=2.3.3", "imageio>=2.37.0", "pillow>=11.3.0", "matplotlib>=3.10.7"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "./src/main.py"
